+++ docs/changelog.rst
@@ -5,6 +5,9 @@ Changelog
 Development
 ===========
 - (Fill this out as you fix issues and develop your features).
+
+Changes in 0.17.0
+=================
 - Fix .only() working improperly after using .count() of the same instance of QuerySet
 - POTENTIAL BREAKING CHANGE: All result fields are now passed, including internal fields (_cls, _id) when using `QuerySet.as_pymongo` #1976
 - Document a BREAKING CHANGE introduced in 0.15.3 and not reported at that time (#1995)